# Optimize array access and add ability to hint array size