… Jint/Native/Object/ObjectConstructor.cs
… Jint/Native/Object/ObjectInstance.cs
… Jint/Native/RegExp/RegExpPrototype.cs
… Jint/Native/String/StringPrototype.cs
… Jint/Runtime/Interop/MethodInfoFunctionInstance.cs
… Jint/Runtime/MruPropertyCache.cs
… Jint/Runtime/MruPropertyCache2.cs