* Made it possible to hint about wanted array size
* Support for dense arrays
* MruPropertyCache is lazy

## ArrayBenchmark

### Baseline

|             Method |   N |       Mean |     Error |    StdDev |     Gen 0 | Allocated |
|------------------- |---- |-----------:|----------:|----------:|----------:|----------:|
|              Slice | 100 |   2.606 ms | 0.0555 ms | 0.0519 ms |  546.8750 |    2.2 MB |
|             Concat | 100 |   2.758 ms | 0.0231 ms | 0.0205 ms |  593.7500 |   2.38 MB |
|            Unshift | 100 | 107.151 ms | 0.3361 ms | 0.2807 ms | 8250.0000 |  33.16 MB |
|               Push | 100 |  33.993 ms | 0.6668 ms | 0.7134 ms | 4000.0000 |  16.13 MB |
|              Index | 100 |  26.968 ms | 0.0660 ms | 0.0585 ms | 2750.0000 |  11.07 MB |
|                Map | 100 |  12.291 ms | 0.0380 ms | 0.0356 ms | 3546.8750 |  14.23 MB |
|              Apply | 100 |   5.936 ms | 0.0071 ms | 0.0063 ms |  742.1875 |   2.97 MB |
| JsonStringifyParse | 100 |  11.847 ms | 0.0464 ms | 0.0434 ms | 2046.8750 |   8.22 MB |

### After sparse array optimization

|             Method |   N |      Mean |     Error |    StdDev |     Gen 0 | Allocated |
|------------------- |---- |----------:|----------:|----------:|----------:|----------:|
|              Slice | 100 |  1.149 ms | 0.0152 ms | 0.0134 ms |  382.8125 |   1.53 MB |
|             Concat | 100 |  1.305 ms | 0.0027 ms | 0.0022 ms |  457.0313 |   1.83 MB |
|            Unshift | 100 | 46.830 ms | 0.0509 ms | 0.0451 ms | 8125.0000 |  32.51 MB |
|               Push | 100 | 33.087 ms | 0.1223 ms | 0.1144 ms | 3937.5000 |  15.81 MB |
|              Index | 100 | 26.781 ms | 0.0412 ms | 0.0344 ms | 2593.7500 |  10.45 MB |
|                Map | 100 |  8.557 ms | 0.0126 ms | 0.0105 ms | 2953.1250 |  11.87 MB |
|              Apply | 100 |  5.607 ms | 0.0047 ms | 0.0037 ms |  718.7500 |    2.9 MB |
| JsonStringifyParse | 100 | 11.445 ms | 0.0234 ms | 0.0219 ms | 1984.3750 |   7.95 MB |

### After introducing dense array

|             Method |   N |        Mean |      Error |     StdDev |     Gen 0 | Allocated |
|------------------- |---- |------------:|-----------:|-----------:|----------:|----------:|
|              Slice | 100 |    935.5 us |   3.443 us |   3.052 us |  336.9141 |   1.35 MB |
|             Concat | 100 |  1,068.9 us |   2.023 us |   1.892 us |  361.3281 |   1.45 MB |
|            Unshift | 100 | 41,124.9 us | 116.609 us | 103.371 us | 8687.5000 |  34.79 MB |
|               Push | 100 | 28,129.1 us |  77.340 us |  72.344 us | 3625.0000 |  14.53 MB |
|              Index | 100 | 25,813.5 us |  84.770 us |  79.294 us | 2500.0000 |  10.07 MB |
|                Map | 100 |  8,320.5 us |  33.465 us |  27.945 us | 2906.2500 |  11.68 MB |
|              Apply | 100 |  1,160.1 us |  11.889 us |  11.121 us |  369.1406 |   1.48 MB |
| JsonStringifyParse | 100 |  6,885.8 us |  18.351 us |  17.166 us | 1671.8750 |   6.71 MB |

## ArrayStressBenchmark

### Baseline

| Method |  N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |      Gen 2 | Allocated |
|------- |--- |--------:|---------:|---------:|------------:|-----------:|-----------:|----------:|
|   Jint | 20 | 2.520 s | 0.0885 s | 0.0050 s | 223750.0000 | 23250.0000 | 11500.0000 | 979.52 MB |

### After sparse array optimization

| Method |  N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 |     Gen 2 | Allocated |
|------- |--- |--------:|---------:|---------:|------------:|-----------:|----------:|----------:|
|   Jint | 20 | 1.492 s | 0.0479 s | 0.0027 s | 180000.0000 | 20000.0000 | 9750.0000 | 805.61 MB |

### After introducing dense array

| Method |  N |    Mean |    Error |   StdDev |       Gen 0 |     Gen 1 |     Gen 2 | Allocated |
|------- |--- |--------:|---------:|---------:|------------:|----------:|----------:|----------:|
|   Jint | 20 | 2.096 s | 0.0750 s | 0.0042 s | 198416.6667 | 8833.3333 | 3833.3333 | 873.57 MB |

## UncacheableExpressionsBenchmark

### Baseline

|    Method |   N |    Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |--------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 1.062 s | 0.0289 s | 0.0406 s | 202687.5000 | 55312.5000 |  903.2 MB |

### After sparse array optimization

|    Method |   N |     Mean |    Error |   StdDev |       Gen 0 |      Gen 1 | Allocated |
|---------- |---- |---------:|---------:|---------:|------------:|-----------:|----------:|
| Benchmark | 500 | 845.1 ms | 3.704 ms | 5.430 ms | 178058.3333 | 45079.1667 | 783.57 MB |

### After introducing dense array

|    Method |   N |     Mean |    Error |   StdDev |       Gen 0 |     Gen 1 | Allocated |
|---------- |---- |---------:|---------:|---------:|------------:|----------:|----------:|
| Benchmark | 500 | 797.6 ms | 2.303 ms | 3.447 ms | 189070.8333 | 2366.6667 | 761.26 MB |