@lahma 2017-12-24
I've implemented @ayende  's suggestions. I also updated the before and after numbers above, now rebased based on current dev.