@lahma 2017-12-27
I think I've done most I can here now, you can see updated numbers above. To proceed with this we need to process other PRs first, which should also give perf boost here too.