@lahma 2017-12-31
I have implemented the dense array thingie. This helped to squeeze some more speed. ArrayStressBenchmark had negative change due to dense array but it's still faster than before. I guess the unrealistic test case just works better with dictionary based backend.